Restructure request/response classes
As there are now additional endpoints that will use them.
