+++ spec/codelocks/request_spec.rb
@@ -1,6 +1,6 @@
 require "spec_helper"
 
-describe Codelocks::NetCode::Request do
+describe Codelocks::Request do
   before do
     Codelocks.base_uri = "http://wobble.com/"
     Codelocks.api_key = "wibble"
@@ -22,19 +22,19 @@ describe Codelocks::NetCode::Request do
 
     before do
       allow(Codelocks.connection).to receive(:get) { response }
-      allow(Codelocks::NetCode::Request).to receive(:default_params) { default_params }
+      allow(Codelocks::Request).to receive(:default_params) { default_params }
     end
 
 
     it "performs a get request" do
       expect(Codelocks.connection).to receive(:get).with(path, all_params)
-      Codelocks::NetCode::Request.create(path, params)
+      Codelocks::Request.create(path, params)
     end
 
     it "returns a response object" do
       expect(
-        Codelocks::NetCode::Request.create(path, params)
-      ).to be_a(Codelocks::NetCode::Response)
+        Codelocks::Request.create(path, params)
+      ).to be_a(Codelocks::Response)
     end
   end
 end
